requirements.txt
Flask
numpy
numpy-stl
dropbox